front: add e2e tests search and filter rolling stock
