+++ front/src/modules/rollingStock/components/RollingStockSelector/SearchRollingStock.tsx
@@ -114,7 +114,7 @@ const SearchRollingStock = ({
         </div>
       </div>
       <div className="col-md-2 mt-1 ml-auto">
-        <small>
+        <small data-testid="search-results-text">
           {filteredRollingStockList.length > 0
             ? `${filteredRollingStockList.length} ${t('resultsFound')}`
             : t('noResultFound')}
