+++ front/tests/005-operational-studies.spec.ts
@@ -4,7 +4,7 @@ import { v4 as uuidv4 } from 'uuid';
 import type { Infra, Project, RollingStock, Scenario, Study } from 'common/api/osrdEditoastApi';
 
 import scenarioData from './assets/operationStudies/scenario.json';
-import { getProject, getStudy, getRollingStock, postApiRequest, getInfra } from './assets/utils';
+import { getProject, getStudy, getRollingStock, postApiRequest, getInfra } from './utils/index';
 import { PlaywrightHomePage } from './pages/home-page-model';
 import RollingStockSelectorPage from './pages/rolling-stock-selector-page';
 import PlaywrightScenarioPage from './pages/scenario-page-model';
